SmurfSpotlight.js work correctly now. Before, it would work if clicked on a smurf from the village, but not if you entered the url directly (/smurf/:id) - now it works.
