+++ village/src/components/SmurfSpotlight.js
@@ -1,24 +1,36 @@
 import React from "react";
 
-const SmurfSpotlight = props => {
-  const id = parseInt(props.match.params.id, 10);
-  const spotlightedSmurf = props.smurfs.find(smurf => id === smurf.id);
+let spotlightedSmurf;
 
-  const deleteSpotlightedSmurf = () => {
-    props.deleteSmurf(spotlightedSmurf.id);
-    props.history.push("/");
+// I don't know if this needs to be a class component or not. It doesn't use state, but I think it needs the render method...
+class SmurfSpotlight extends React.Component {
+  deleteSpotlightedSmurf = () => {
+    this.props.deleteSmurf(spotlightedSmurf.id);
+    this.props.history.push("/");
   };
 
-  return (
-    <div className="smurf spotlight">
-      <h2>{spotlightedSmurf.name}</h2>
-      <strong>{spotlightedSmurf.height} tall</strong>
-      <p>{spotlightedSmurf.age} smurf years old</p>
-      <button onClick={deleteSpotlightedSmurf} className="delete-smurf-button">
-        Delete Smurf
-      </button>
-    </div>
-  );
-};
+  render() {
+    const id = parseInt(this.props.match.params.id, 10);
+    spotlightedSmurf = this.props.smurfs.find(smurf => id === smurf.id);
+
+    if (!spotlightedSmurf) {
+      return <h3>Loading smurf...</h3>;
+    }
+
+    return (
+      <div className="smurf spotlight">
+        <h2>{spotlightedSmurf.name}</h2>
+        <strong>{spotlightedSmurf.height} tall</strong>
+        <p>{spotlightedSmurf.age} smurf years old</p>
+        <button
+          onClick={this.deleteSpotlightedSmurf}
+          className="delete-smurf-button"
+        >
+          Delete Smurf
+        </button>
+      </div>
+    );
+  }
+}
 
 export default SmurfSpotlight;
